Add files via upload
osc_energy.m is a function to compute energy of an oscillating system.
osc_energy_EC.m is a program using Euler-Cromer scheme to solve the dynamic model of an oscillating system. Displays of magnitude and total energy.
osc_energy_FE.m is a program using Forward-Euler scheme to solve the dynamic model of an oscillating system. Displays of magnitude and total energy.